Adding removal of non user tub data to donkey tubcheck
Add exception to get_json_record() when "user/model" is not "user". This indicates the data is from an autopilot and will have 0 "user/angle" so will corrupt the tub data for training. Also corrected exception text in same function.
Corrected help text for donkey tubcheck.
Removed python script with the same functionality.